tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/actuator/IrsConnectionHealthIndicator.java
/********************************************************************************
 * Copyright (c) 2023 Contributors to the Eclipse Foundation
 *
 * See the NOTICE file(s) distributed with this work for additional
 * information regarding copyright ownership.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Apache License, Version 2.0 which is available at
 * https://www.apache.org/licenses/LICENSE-2.0.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations
 * under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 ********************************************************************************/
package org.eclipse.tractusx.traceability.common.actuator;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.IRSApiClient;
import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;
import org.springframework.boot.actuate.health.Health;
import org.springframework.boot.actuate.health.HealthIndicator;
import org.springframework.stereotype.Component;

@Component
@Slf4j
@RequiredArgsConstructor
@ConditionalOnEnabledHealthIndicator("irs-connection")
public class IrsConnectionHealthIndicator implements HealthIndicator {

    private final IRSApiClient irsClient;

    @Override
    public Health health() {
        try {
            irsClient.getPolicies();
            return Health.up().build();
        } catch (final Exception e) {
            log.warn("IRS connection health check failed, error: {}", e.getMessage());
            log.debug("IRS connection check failure exception: ", e);
        }
        return Health.outOfService().build();
    }
}
